ssaffice-PY/requirements.txt
APScheduler==3.10.4
boto3==1.35.55
elasticsearch==8.16.0
langchain==0.3.9
langchain_community==0.3.8
langchain_openai==0.2.10
openai==1.55.3
Path==17.0.0
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.33
websocket_client==1.8.0
pydantic==2.10.2
fastapi==0.115.5
uvicorn==0.27.1
PyMySQL==1.1.0